DROID-2121 move search field to core ui
